requirements.txt
annotated-types==0.7.0
awscli==1.44.12
boto3==1.42.22
botocore==1.42.22
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
docutils==0.19
dotenv==0.9.9
filelock==3.20.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
pyasn1==0.6.1
pyclean==3.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.7.2
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
sentry-sdk==2.48.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
wandb==0.23.1
onnx
onnxruntime
dotenv
